+++ tests/testthat/test_hill_func.R
@@ -7,27 +7,27 @@ test_that("FDis, FD vs hillR", {
   expect_equal(a["FDis", ], b)
 })
 
-test_that("functional dissimilarity should be between 0 and 1; q = 0", {
+test_that("functional similarity should be between 0 and 1; q = 0", {
   a = hillR::hill_func_parti(comm = dummy$abun, traits = dummy$trait, q = 0)
-  expect_lte(a$local_dissimilarity, 1)
-  expect_gte(a$local_dissimilarity, 0)
-  expect_lte(a$region_dissimilarity, 1)
-  expect_gte(a$region_dissimilarity, 0)
+  expect_lte(a$local_similarity, 1)
+  expect_gte(a$local_similarity, 0)
+  expect_lte(a$region_similarity, 1)
+  expect_gte(a$region_similarity, 0)
 })
 
-test_that("functional dissimilarity should be between 0 and 1; q = 1", {
+test_that("functional similarity should be between 0 and 1; q = 1", {
   a = hillR::hill_func_parti(comm = dummy$abun, traits = dummy$trait, q = 1)
-  expect_lte(a$local_dissimilarity, 1)
-  expect_gte(a$local_dissimilarity, 0)
-  expect_lte(a$region_dissimilarity, 1)
-  expect_gte(a$region_dissimilarity, 0)
+  expect_lte(a$local_similarity, 1)
+  expect_gte(a$local_similarity, 0)
+  expect_lte(a$region_similarity, 1)
+  expect_gte(a$region_similarity, 0)
 })
 
-test_that("functional dissimilarity should be between 0 and 1; q = 2", {
+test_that("functional similarity should be between 0 and 1; q = 2", {
   a = hillR::hill_func_parti(comm = dummy$abun, traits = dummy$trait, q = 2)
-  expect_lte(a$local_dissimilarity, 1)
-  expect_gte(a$local_dissimilarity, 0)
-  expect_lte(a$region_dissimilarity, 1)
-  expect_gte(a$region_dissimilarity, 0)
+  expect_lte(a$local_similarity, 1)
+  expect_gte(a$local_similarity, 0)
+  expect_lte(a$region_similarity, 1)
+  expect_gte(a$region_similarity, 0)
 })