+++ NEWS.md
@@ -4,3 +4,4 @@
 - `hill_func` and `hill_func_parti` no longer standardize distance matrix to have max value of 1 by default.
   + set `stand_dij = TRUE` if want to do so.
 - `hill_func_parti` function now reports site **similarity** instead of dissimilarity, to be consistent with the `hill_taxa_parti` functions. Make sure to change your interpretations.
+- Added `hill_phylo` functions.